requirements.txt
asgiref==3.11.1
asttokens==3.0.1
decorator==5.2.1
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-jalali-date==1.0.2
django-picklefield==3.1
django-q2==1.6.2
django-rest-passwordreset==1.4.0
django-unfold==0.22.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.2.1
gunicorn==21.2.0
ipython==8.20.0
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.19.2
matplotlib-inline==0.2.1
mysqlclient==2.2.1
packaging==26.0
parso==0.8.5
pexpect==4.9.0
pillow==10.2.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.8.2
python-decouple==3.8
python-magic==0.4.27
pytz==2025.2
redis==5.0.1
setuptools==81.0.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.6.0
whitenoise==6.6.0
